# server/docs/swagger.yaml
definitions:
...
  models.ExpenseReceiptRef:
    properties:
      _id:
        type: string
      contentType:
        type: string
      height:
        type: integer
      uploadedAt:
        type: integer
      uploadedBy:
        type: string
      width:
        type: integer
    type: object
...
  responses.Error:
    properties:
      error: {}
    required:
    - error
    type: object
...
host: localhost:3002/api
info:
  contact: {}
  description: This is the API for The Fellowship!
  title: The Fellowship API
  version: "1.0"
paths:
...
  /events/{eventId}/expenses/{expenseId}/receipts:
    post:
      consumes:
      - application/json
      description: Accepts a base64 data URL (JPEG or PNG). The image is downscaled
        to a 2000px long edge and re-encoded as JPEG before storage, so EXIF metadata
        — including the GPS tag on a phone photo — is discarded. The member who entered
        the expense, or any admin.
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      - description: Expense ID
        in: path
        name: expenseId
        required: true
        type: string
      - description: Data URL of the image to store
        in: body
        name: payload
        required: true
        schema:
          properties:
            image:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ExpenseReceiptRef'
        "400":
          description: invalid-image
          schema:
            $ref: '#/definitions/responses.Error'
        "413":
          description: image-too-large / too-many-receipts
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Attaches a receipt photo to an expense
      tags:
      - events
  /events/{eventId}/expenses/{expenseId}/receipts/{receiptId}:
    delete:
      description: The member who entered the expense, or any admin.
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      - description: Expense ID
        in: path
        name: expenseId
        required: true
        type: string
      - description: Receipt ID
        in: path
        name: receiptId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: not-authorized
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Removes a receipt photo from an expense
      tags:
      - events
    get:
      description: Readable by anyone signed in who can see the gathering, guests
        included — the ledger is theirs to read too.
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      - description: Expense ID
        in: path
        name: expenseId
        required: true
        type: string
      - description: Receipt ID
        in: path
        name: receiptId
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
        "404":
          description: receipt-not-found
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Serves a receipt photo
      tags:
      - events
...
  /user/avatar:
    delete:
      description: Idempotent — removing a photo that isn't there succeeds.
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Removes the signed-in user's profile photo
      tags:
      - user
    put:
      consumes:
      - application/json
      description: |-
        Accepts a base64 data URL (JPEG or PNG). The image is centre-cropped,
        downscaled to 256x256 and re-encoded as JPEG before storage, so EXIF
        metadata on the original is discarded.
      parameters:
      - description: Data URL of the image to store
        in: body
        name: payload
        required: true
        schema:
          properties:
            image:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              avatarUpdatedAt:
                type: integer
            type: object
        "400":
          description: invalid-image
          schema:
            $ref: '#/definitions/responses.Error'
        "413":
          description: image-too-large
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Sets the signed-in user's profile photo
      tags:
      - user
...
swagger: "2.0"